B18-520 : Make fluorescence detector view use max detector counts setting
Each time a fluorescence detector view is opened, the client config is
searched to try and locate a view factory object for the detector. If
one is found, the max counts and MCA scannables settings from it are
applied to the view.
This ensures the same settings are used in the view for a detector,
whether it is opened from 'show view' menu (or equivalent), or from
the 'detector parameters' GUI.

Change-Id: Ie14403f4e84098aee68d52e6554ab117ee11ccb3