B18-523 : Make Fluo detector view work with odd and non-square number of detector elements
The Detector element table in the 'fluorescence detector' view now works
correctly when there is a non square or odd odd number of detector elements.
- The Richbeans GridListEditor that used to display the detector elements
is unable to cope with partially filled grid of detector elements. It has
been replaced by a grid layout of Labels, with appropriate mouse listeners
to update the colours and notify observers.
- The import ROI wizard has been updated to take the number of detector
elements directly, rather than extract the number from the GridListEditor
of detector elements.

Change-Id: I6faf7189d0b568330c27652cd45ba86240eb745f