B18-522 : Add DTC factor calculation to GDA for Xspress4
Xspress4BufferedDetector can now calculate the deadtime correction (DTC)
factors from the raw scaler values. This is to overcome a limitation
with
current Xspress IOCs - DTC factor values are not available in the Hdf
file
or as a time series array PV, and so are not available during continuous
scans.
- Old Xspress2 code for dealing with DTC factor calculations has been
re-used.
- Deadtime correction parameters can be stored in an XML file as an
XspressDeadTimeParameters bean and can be loaded into Xspress4Detector.
The parameters are then available to Xspress4BufferedDetector as needed.
- GDA calculation of DTC factors during scans can be enabled/disabled
using the Xspress4BufferedDetector#calculateDtcFactors field.
- Unit tests have been added to ensure that GDA calculations match DTC
factors produced by Epics during step scans.

Change-Id: I7d691de09db5235d4e68bc227a966139b5c85a01