Add new options to gda-conf.py to make it more user friendly
Add a [descriptions] section to the config file to provide a description
of each device. Descriptions with the same key as a device are shown
along side each device. Missing and extra descriptions are ignored.

If the tooltips option is used, descriptions are added as tooltips on
the device checkbox rather than being given their own column.

Also,
* Default to using gda_versions/var/BEAMLINE.cfg as the config location
  if no location is specified.
* Add command line and config file options for prefix and suffix text,
  to provide help and/or instructions to users.
* Add command line and config file options for setting the font.
* Add a tooltip for where selections will be saved.
* Add help text for existing command line options.
* Add an example of the new config file options.

Change-Id: Ia6f4ebed9475f5c2bef7befab1361067a1f5fdbc