Make BeamMonitor implement existing IBeamMonitor
To allow it to be used in place of the older IonChamberBeamMonitor.
The waitForBeam method has been moved into the BeamMonitor class
instead of in the base beam condition class. This allows it to
break out of the wait method if the monitor is disabled why the
beam is down.

Change-Id: Ief57e89b02d37a1688f897081d193448ff5a3bfa