K11-1694 Abort running plan when 'stop experiment' button clicked
Now the plan subscribes to the experiment controller so that should the
experiment end (because the user clicks the stop button in the GUI), the
plan will abort. Previous behaviour remains: if a reference to the plan
is held (say, for a script-based experiment), abort() can be called and
ExperimentController::stopExperiment is invoked as well.

Other changes include
 - Document declaration moved from bean (impl) to PlanRequest
 - Anticipating other final states than COMPLETE in listening GUI parts
   (TERMINATED when plan is aborted)
 - Test fixes

Change-Id: I17d0aef9b0a9ce20f8c140175d07f381388b8a13