I14 Add check for decay mode to topup watchdog to pause/resume scans
Currently scans are paused when a top up event happens. The topup
countdown value stays in -1 during the top up filling period. However,
when the beam is in 'decay mode', the value stays at -1 and the scan
remains paused even though there is no top up event happening.

Decay mode depends on the value of the PV: SR-CS-FILL-01:TOPUPSTATE
If top up state is in 'Failed', 'Off' or 'No process' then the beam is
in decay mode.

I14 wants to continue scanning while the beam is in decay mode.

With this new change, if:

- If toup_countdown < 0 && decayMode is False then Scan Status =  PAUSED
If decayMode is True then Scan Status = RESUMED

- If toup_countdown > 0 && decayMode is False && Scan Status = RUNNING
If topup_countdown is < 0 then Scan Status = RUNNING (It is not paused)


Change-Id: I82481c2aa70b5548a1d2657ace7164d495b5aaee