I14-962 Add custom converter to allow 5 decimal places
Default data binding context converters from double to string are set to
round the exposure time to the nearest 3 decimal places.

Added a new custom NumberToStringConverter to round the number to 5
decimal places instead as I14 would normally set the Eiger detector
exposure time to 0.00075.

The IObservableValue old implementation without declaring a type had to
be used to make the data binding with the custom decimalConverter to
work. Warnings are suppressed.


Change-Id: I6c6af9351db30534f07e36310e4f01caaba8de4c