I14-966 Add the "sparse scanning" XANES acquisition to GDA
Changes to support adding "sparse scanning XANES acquisition" as a
ScanType

- DetectorsSection:

New method that finds a detector by name and adds it to the list of
visible detectors if it is not already there. It selects/deselects a
detector checkbox according to a boolean and notifies the listenres to
trigger the method detectorSelectionChanged which would normally do if
this checkbox would be clicked manually.

- SubmitScanToScriptSection

New method that calls the previous explained method selectDetector in
DetectorSection. This is almost identical to selectOuterScannable but
with a detector. The mappingView.updateControls is not called as this
enables other detectors that were previously disabled as a malcolm
detector has been selected.

- XanesEdgeParameteres

New parameter called percentage that defines the percentage of rows to
keep from the defined region

- XanesEdgeParametersSection

Added a new spinner that lets the user select the percentage of y
positions to scan.
Added two methods to show/hide two composites: LinesToTrack and
Percentage depending on the type of Xanes Scan is running. In the
future, I14 is planning to use tracking for Sparse Xanes.

- XanesSubmitScanSection

Added getters to support the configuration of buttonColour,
sparseSpanning, scriptFilePath and detectorName.

onShow() method selects a detector if a a detectorName was configured,
and onHide() deselects it when the view is disposed

Depending on the type of xanes scans, the percentage and linesToTrack
composites are visible or not

Change-Id: Ida02eb9f50cc1d9af94bdd59678ff746ae2eec41