Switch Double MIN & MAX VALUE for NEGATIVE/POSITIVE INFINITY where needed
Double.MIN_VALUE is close to zero, so using it for a lower limit for a
positioner, or the starting point for a min/max loop could result in
invalid limits and mix/max results.

Where the use of Double.MIN_VALUEis valid, but inobvious, add a comment
to detail bring attention to that.

Also identify other potential uses of Double.MIN_VALUE where the apparent
intention does not match with it's use.

Change-Id: I19e5b324bf83df77a572ee597c79b5353f4aa253