Fix plan parsing when SEV is driver readout
Recently all scannables becames available for SEV selection in the plan
setup UI, but this broke previous behaviour where SEV was a driver
readout e.g. tr6's load.

The fix: we check whether a driver is included in the plan and if it is,
whether the SEV name returns anything meaningful when treated as a
driver readout. If not, assuming the SEV is any scannable in the GDA
configuration, we retrieve it through the Finder.

Change-Id: Iee03604a61d778142961fa236f1c07195f3cf252