Remove outdated (and now buggy) logic in TriggerEditor
The logic in toggleSourceSelection() may have been useful when only
driver readouts were valid SEV options, but that is no longer the case.

These calls actually resulted in a bug when loading a plan with driver
readout-based triggers, where the selection would flip to time-based.

The existing binding logic is sufficient to set the initial state of
widgets, so this method can be removed.

Change-Id: Icac0a0809d301edb9b63636f993a2e43063d0245