I18-551 Don't translate attribute access on aliases
If a method was called on an aliased name, it would be translated as if
the attribute name was the first argument passed to the function.

eg
    >>> alias('foo')
    >>> foo.bar()
        File "<input>", line 1
            foo(.bar())
                ^
    SyntaxError: no viable alternative at input '.'

This adds an additional check so that aliased commands are only
translated if they are followed by whitespace. This also allows indexing
etc on aliased objects.

Change-Id: I8bf24b195732737c52ddd4820e516024dc1a642f