Refresh configuration when creating SingleFileAuthoriser
The XMLConfiguration instances returned by LocalParameters are cached so
changes made to the file are not automatically picked up from one access
to the next. For SingleFileAuthoriser this meant that when the
permissions file was modified, the servers had to be restarted for the
changes to have any effect even though a new instance was created for
each client.

Refreshing the configuration means that only the client needs to be
restarted to change permissions for a user.

Change-Id: If8ffd07f035044c55c8330a585bb97b03e5b92e5