I14-932 Add exposure controls to CameraConfiguration
This change includes:
- Adding a CameraControlsComposite which instantiate a
LiveStreamViewCameraControlsScanListenerwhich enables/disables controls
to prevent users from starting/stopping
a detector acquisition while a scan is running.

- When a detector is already acquiring and the boolean
changeExposureWhileCameraAcquiring is false, if a user attempts to
change the exposure, the text label would be misleading as it would show
the last entered value but that it is not the real exposure time set.
This now sets again the current exposure time if the attempt to set it
was unsuccessful.

- Increased the width of the exposure label to allow more decimal
places.

- CoameraConfigurationTabs only create tabs depending on the detectors'
configuration. There is a new listener for detector changes, so there is
no need for other listeners in the tabs composites.

Change-Id: I0c264ab3fe07d53de09128008ca1ef69d7507787