Document org.eclipse.scanning.device.Transformation.setType for Spring
It is not obvious how to use setType(TransformationType type) property
in Spring, so document what the spring property needs to look like.

Also add javadoc for the TransformationType enum and deprecate the
setType(String typeStr), since it is neither called from Java, nor the
Spring framework.

Attempting to use

	<property name="type" value="rotation" />

results in setType(TransformationType type) being called and thus the
error:

Cannot convert value of type 'java.lang.String' to required type
'org.eclipse.scanning.device.Transformation$TransformationType' for
property 'type': no matching editors or conversion strategy found

We see the same result if we explicitly do

	<property name="type">
		<value type="java.lang.String">rotation</value>
	</property>

Since these errors do not make the clear what the problem is, or how to
solve it, do so in the documentation, using the variant name rather than
the toString() value:

	<property name="type" value="ROTATION" />

Change-Id: I439d5a3449934e807eddc36c1f54a1b43b21bca7