DAQ-4344 Delay OSGi service registration in SpringProxyParser
If the service is registered when the bean is parsed, it requires the
RMI connection to be available so cannot be used in test contexts. This
change delays the registration by creating the OSGiServiceRegister beans
that would be created manually if this parser wasn't being used. The
registration is then handled at the 'afterPropertiesSet' stage after
instances have been created.

These new beans are named "<bean_id>_<service_class>". If these are
found to cause naming collisions, a more robust naming scheme can be
used but this should be unlikely.

Change-Id: I6ef39c7ca0f7524b3e3e492e06089aec5dae8dfa