I20-567 : Disable 'Two colour' option in GUI if it is not allowed for scan type
- 'Two colour' option is disabled in 'scan colour' radio buttons if it
is
not allowed for the currently selected scan type; Two colour' scans are
currently only allowed for 'fixed XES scan XAS/XANES' scans.
- Workaround for Richbeans bug : NumberBox widgets throw a NPE if a
value
is set on a disabled widget. Workaround is to temporarily enable the
widget,
set the value, and restore the original enabled/disabled state.
- Added a dispose listener in SignalParamtetersComposite to dispose of
leftover images in a RegularExpressionTextWrapper widget.

Change-Id: If568bab7bee531e7ae8062ffea8dc84daf3a6c3e