I16-7722 (9.28) Fix duplicate secondaryId in LiveStreamMenuContribution
If CAM1#MJPEG already existed when trying to open CAM1#EPICS_ARRAY then
the menu action tried to open a view with an effective secondary ID of
CAM1#MJPEG:CAM1#EPICS_ARRAY which failed.

Instead of searching for existing views which match just CAM1, search
for existing views which match the desired secondaryId, for example
CAM1#EPICS_ARRAY, and then either set secondaryId to null if one is
found (since it already contains the required secondaryId), or create a
new view with the desired secondaryId if it isn't found.

Change-Id: Ib653c2238d70c83c937d867c5b06e454f4e69a3f