I20-569 : Update calculations for XES spectrometer
Updates to XesEnergyScannable :
- Functions have been added to XesEnergyScannable to return the
positions of all scannables in the spectrometer for a given energy. This
removes the need to have a dummy version of the spectrometer to
calculate the motor positions.
- Calculation has been updated to use the documented set of equations
for crystal position and orientation.
- Absolute X position of each crystal can be used by setting
absoluteXPos to true (this is required for the new 14 crystal
spectrometer)
- Detector translation axis can now be at an angle w.r.t. the analyser
coordinate system.
- Offset calculation has been updated to use the new position
calculation functions.
- Unit tests have been updated to use the new position calculation
functions.

Change-Id: I92bd17949cb4ea3100e47a652cb541040a518224