I14-951 Enable lines to track selection
Lines to track used to be set by selecting a "dawn xrf processing" file.

Beamline staff has requested to select element-to-align manually in gda.

Combo box is now separated into two combobox viewers, one for the
element and the other one for the line macro.

The line macros list is updated depending on the element selected
according to xraylib.

There is an empty space in the list of elements in case they don't want
to run xanes with tracking.

Change-Id: I1b55b9a4a4c47a4d6f7e4ded8fe27ddae95dc585