I14-991 Remove malcolm detectors marked as offline
If a Malcolm detector is offline, its label has an * to indicate that
this detector is not available.

When the Malcolm detector is online again and gda client is restarted
(without clearing) the detector parameters in the bean still have the
malcolm detectors with the *. Then, there are duplicated detector names,
one with the * and one without *. For example: "Simulated motors for
mapping" and, "Simulated motors for mapping [*]"

Because the detector is working again, the one with the * should not be
displayed.

This change removes the detectors that are online but have * in their
label.

Change-Id: I3dcf3e0866ef6b67d655fbd228ae8140428ffd0c