[DAQ-4318] write axes attribute correctly for flattened grids
Previously a data device was either a default axis device for a
particular dimension of a scan, or was not an axis at all. In a
flattened grid, both x and y axes were added as default axis, so
one was overwritten in the @axes attribute by the other, and in an
unpredictable way.

Now, only the first axes for a dimension is the default. However, we
still want to add the other axes as axes, so we need an extra boolean
property 'defaultAxes' to distinguish between default axes and
non-default axes.

Change-Id: Ifdeaf0d45c2d3bfb04b25d7102d65d67600b8959