K11-1742 Refactor ScanpathDocument
ScanpathDocument and ScannableTrackDocument did not readily accommodate
diffraction tomography scans. ScanpathDocument consisted of a list of
ScannableTrackDocuments with a single field hinting the 'shape' i.e. the
relationship between each scannable. This is insufficient: Diff tomo
is essentially defined as a 2D diffraction whose shape may vary, with a
1D outer scan. Further, the execution of this scan needs to happen in a
different order (X vs theta vs Y). This is currently not possible; the
ScanpathDocument gets converted to a single AcquisitionTemplate which
generates scan point generator and ROI to send to Malcolm.

Now, a ScanpathDocument describes a collection of N-dimensional trajec-
tories. Each trajectory then produces its own SPG and ROI contributing
to the CompoundModel sent to Malcolm. This can easily describe the new
scan type and future dimensions.

Summary of the changes:
- ScanpathDocument now consists solely of a list of trajectories
- Trajectories encapsulate a list of axes (ScannableTrackDocument) and
  a shape describing their relationship
- Default trajectories specified in XML-configured AcquisitionTemplate
  (the one in uk.ac.gda.client.properties.acquisition)
- An earlier change making AcquisitionTemplate return multiple SPGs is
  reverted
- Instead, AcquisitionTemplateFactory returns multiple instances of
  AcquisitionTemplate from a single ScanpathDocument (K11-1746)
- AcquisitionTemplateType renamed to TrajectoryShape and deleted from
  this enumeration DIFFRACTION_TOMOGRAPHY (K11-1744)
- Brought ScanningParametersUtils from gda-diamond.git and added extra
  utils
- Deleted pointless builder for ScanpathDocument

Change-Id: I8541e1a8b762b77a2c433590dadc2e113131d5b8