I16-664 (9.28) Restore determinism to EpicsControlPoint.setValue()
Restore synchronous behaviour to EpicsControlPoint.setValue(), while
keeping asynchronous behaviour for asynchronousMoveTo().

Also, fix a race condition in EpicsControlPoint.asynchronousMoveTo()

Using Async.execute, it was highly likely that the calling thread would
be competed first, so doing `asynchronousMoveTo(1); waitWhileBusy();
asynchronousMoveTo(2); a=getValue()` would result in `a` having the
value of `1`, rather than the expected value of `2` since the
`getValue()` call would be made on the original thread before the thread
scheduled by `setValue(2)` had a chance to run.

To remove this race condition, do the caput immediately, but use a
PutListener to allow it to return straight away, rather than scheduling
teh caput to happen later in a thread.

This problem was discovered and fixed on the beamline after the contents
of gda-epics.git were moved into gda-core.git so this change is a
re-implementation of that unmergable commit.

Change-Id: I4d7f657d3b396195382488edb3d60df96fb838ee