DAQ-4467 Restructure stdin handling in Jython
Although the InputCommands methods are not required in Jython when
running via SSH, it should be possible to use them without the shell
hanging.

This moves the input handling from the InteractiveConsole into the
JythonServerThread if one is running the command and falls back to
the RCP input if not. As the input stream is available in the Jython
thread, all methods of requesting input can now handle any Jython
context, including SSH.

Change-Id: I64f54e180a8a1eed6127298e1e48d389fbaee3e7