K11-1754 Add acquire button
The 'Controls' tab which shows exposure and binning controls now also
creates a button to acquire a snapshot to file (and a box showing the
filepath) *if* the CameraConfigurationProperties has a non-null acqui-
sition device name. This name is the name of the corresponding runnable
device, to be retrieved via runnable device service.

Change-Id: I5b900ae77074a9aeb7a60f43d688b0fc8b982613