Fix ADPilatusTest
This test previously was failing however this was not detected as the
AssertionError was thrown in a different thread.

This is resolved by running the task with modern asynchronous Java.

The test was also modified to pass as the current behaviour of ADPilatus
is used on beamlines and supposedly working as expected.

Change-Id: I08fdb72b1db3fefa15dc4058de131afb06d73549