K11-1705 Make sure latest acquisition per acquisition key is stored
AcquisitionManager has a map of acquisitions, but only inserted items
in there when getAcquisition(key) was called for each new key. Now
internally caches acquisitions created with newAcquisition(key), and
for cases when acquisitions are obtained externally e.g. loading from
file, they can now be registered with the manager.

Change-Id: Iff40d5259f6919b6081969bc97da53fcbd4585ac