Improve handling of Jython help function
Rename `_gdahelp` to `help`. The GeneralTranslator used to have a
hardcoded alias to map help commands to _gdahelp to better handle
objects derived from Java classes. Renaming the help method means the
redirection is not required and new Translators can still access the
custom help.

Replace Java type checks. Previously, a class that extended a Java type
would not have any custom help displayed as its own `__class__` (`type`)
would not be a subclass of java.lang.Object and it would fall back to
the very limited auto generated help. The new method checks for either
extending java Object or being a type that is a subclass of Object
allowing both classes and instances of those classes to have useful help
shown if available.

Change-Id: I373b539c6c9e86ad500c309cf1b437e3cf9aa32f