I20-575 : Updates for XES spectrometer after beamline testing
- Detector x, y and angle values are now multiplied by the x, y and
pitch multipliers, the same as for the analyser stages.
- The current Bragg angle value is now derived from the current
angle of the the detector. This avoids problems caused by using the
detector y value which may be in a rotated coordinate system.
- Y and pitch values are now reversed for the *lower' set of analyser
stages, not the upper.
- The energy and angle in the 'XES Simulated positions' view now work
more reliably, by setting the initial energy and angle directly on the
dummy motor objects. Also, the energy/angle update observers are now
removed when the view closes.

Change-Id: Ib044f407e9032d8eead0367595a6d7a9290290c7