I20-579 : Updates to get XES scans working with new colour options
Updates when using a List of DetectorConfigs to define detector
configuration for a XAS scan :
- XasScan will now automatically generate list of detector objects
to be used from the selected detectors in the list.
- ExafsValidators will now check each of the Fluorescence/Medipix
detector exists and that : the number of detector elements is consistent
with the detector object (Fluorescence detector), 1 or more ROI has
been set (Medipix).
- DetectorPreparerFunctions can now be used with multiple different
medipix detectors; ROI plugins for each are stored in a map.

Change-Id: I5daba213d0ef0d5b7c98caf64b23b01bb793f9ae