Let IRerunHandler operate on multiple scans
The signature of IRerunHandler.run is changed to take a list of scans
instead of a single one. Note there are currently no implementations of
this interface.

StatusQueueView has a default implementation of rerun, which is here
also changed to operate on the entire selection. In practice this means
users get a single confirmation dialog instead of one per selected scan.

Finally, this change also fixes some minor compiler/Sonar warnings.

Change-Id: I48b6ace625e521e0eb6fb28d6e5ed14bdc3c2b2e