K11-1767 Remove ExperimentTreeCache
This object serialised to file the experiment state every time it
changed so that it could be restored when the server was restarted.

This feature is no longer wanted: beamline staff want to be able to
reset most things by restarting the server, especially when they cannot
get hold of the technical support team.

Some logging at debug level was added.

Change-Id: Ie5e284a9dc4be7566c784e3756b19979037b8b47