remove ScanDataPointUtils and replace with generic method in AbstractCachedScanPlotView
each method in ScanDataPointUtils repeated the same logic for a different
header name. The new method in AbstractCachedScanPlotView removes the
duplication and makese use of strings. It has been moved to that class
as it is the abstract superclass of the only two classes that actually
made use of ScanDataPointUtils.

Additionally, the logic of the updateCache methods in each of the two
classes, ExafsScanPlotView and LnI0ItScanPlotView has been simplified,
removing unnecessary use of the keyword 'continue'.

Change-Id: I8a047c58448075889a16762b08433da28312d7f1