Add new configuration loading
The new configuration loading framework is designed to be modular and
allow configuration to be provided by multiple independent sources.
These are then combined in a service providing a single access point for
all GDA components requiring configuration options.

The three sources currently implemented are CLI options, environment
variables and configuration directories.

CLI options can be provided for the core options for spring xml,
property files and logging configuration. Any additional, unrecognised
options are set as properties to be combined with those read from the
properties files. - See `CliOptions.java`

Environment variables are filtered to only use those with a `GDA_`
prefix, and then they they are set as properties to be combined with the
loaded properties. The core configuration elements are taken from
predefined variable names. - See `GdaEnvironment.java`

Configuration directories are handled differently from the convention
based layout approach used previously. It is now more declarative with a
`config.toml` file expected to be in the root of the directory. This
file lists which files should be used for each of the core options. It
also allows additional directories to be included allowing common
configuration components to be shared by multiple deployments. This
approach is intended to make it clearer which files are being included
and why without having to read through multiple xml files following
import elements.

As config directories are more complex, their implementation is spread
over multiple classes in the
uk.ac.diamond.daq.server.configuration.source.directory package.

Change-Id: I77984a87163696db0538bd592c8c2b39117a625f