Add ability to configure logging with multiple files
And pass in properties at runtime rather than requiring them to be
imported from files linked from the config files.

Logback is designed to be configured by a single file (that can
optionally import additional files). This means that if multiple files
with top level `Configuration` elements are used, its internal tracking
of which files are used gets out of sync with the files that are
actually being used and the default 'scan' mechanism can't be used.

It would also be more consistent to use `include` as the top level
elements as all files used are equal and there is no concept of a root
file. Any 'scan', 'debug' or 'scanPeriod' attributes added to
`Configuration` elements would be ignored which could be misleading when
debugging any issues.

For this reason, we treat all files as being included by a virtual top
level file that only contains includes. This file doesn't exist -
instead the sax parsing events that would be generated by parsing such a
file are passed to the configurator so it behaves as if it did exist.
The scan/auto-reload behaviour is reimplemented so that the same process
can be used when the logger context is reset. This relies on Java's
built in FileWatcher mechanism and the logger context's own scheduled
executor service. Each monitor task runs repeatedly until a file is
changed at which point it is cancelled and replaced by a new task that's
created by the method that refreshes the context.

For now, the refresh period is hardcoded to 10 seconds (the previous
default), but if a different time would be useful in some circumstances
it could be made configurable.

Change-Id: Ic2e1be88e1348abfcfa9f7ad6779505c5d3abd59