DAQ-4547 Stop caching builtins in completer
If builtins are added dynamically after the jython completer has been
created, they are never made available as options when calling
auto-complete.

While there is a slight overhead in computing the builtin completions
each time, it is minimal when compared to the work done computing the
matching globals. If this does end up being a noticeable regression, we
can look at implementing some way of refreshing the autocompletion and
rebuilding the cache.

Change-Id: I0575ddca022c05eb02df9e7d8261d30acd03f27c