Ignore metadata scannables with a null position in NexusDataWriter
This is the same behaviour as NexusScanDataWriter, will not change the
behaviour of any beamline with a working NexusDataWriter configuration,
but will allow a NexusDataWriter configuration requiring context
dependent metadata to work.

For example a SimpleScannable, added as a metadata scannable, which is
saved into the Nexus file if it has a value, but not saved if it
doesn't, preventing the nexus file being cluttered with an irrelevant
empty node.

Change-Id: I9089c8156dce573ec1efac1e5af9089da5c21558