I20-579 : Allow XasAsciiDataWriter column ordering to be customised
The methods used to control the order and name of the data written by
XasAsciiDataWriter have been rationalised :
- XasScanDataPointFormatter now contains a map which controls the order
the data is added and the name of the column. The default map (for XAS
scans) gives the same behaviour as before. An alternative/additional map
can also be passed in to override/customise the behaviour. These
settings are passed via the AsciiDataWriterConfiguration object.
- XasAsciiDataWriter#replaceDetectorHeaderNames has been removed, since
this functionality is now handled in XasScanDataPointFormatter.

Change-Id: I264519f2f841fe0cd235d35fecf389b1c0d07d1e