Add CoreConfigurationSource and register server/client services
The CoreConfigurationSource is similar to the ConfigurationSource
interface used in the configuration framework but does not allow
implementations to overwrite other sources. It also provides no method
of setting properties directly as any properties are expected to be
fixed and can therefore be included in the files. Paths returned for the
server/client/properties etc can still include properties used for
variable expansion (to allow different files to be used in different
modes etc).

The only current implementation is to load files from a
`resource/core-config` directory at the root level of a plugin. If this
implementation is then set as a service, the root configuration service
can provide configuration from multiple plugin sources without directly
depending on them.

This is done for both the client and the server to configure core
components required for each (logging, properties and spring) as well as
components common to both in the plugin containing the config loading
framework itself.

Change-Id: I361038949089e4ae5b06effadc4aaad74133a13c