Provide core configuration from built product sources
The CoreConfigurationSources defined in the server and client plugins
are now included as additional sources of configuration options in the
configuration service. These options are fixed and will always be
present even when the existing sources specify that they don't use
default options. This is intentional and prevents core components being
replaced or removed accidentally.

Change-Id: I1f710752ba77d6ca4f55f281946459e1e796115e