I14 Fix unselected tracking input and tracking method radio button
When tracking was unselected, LineToTrack was still set to the previous
value (eg Ni-La) even though in the gui this was empty. Now, when
tracking is unselected, the value is set to null for running scans with
no tracking.

When the radio buttons were selected programatically, the model would
not be notified of the new selection. Now, the buttons notify the
listeners.

Change-Id: I9d8c909fa5a5ccd23f20e5b7c0183a70f97e9c28