Ensure pressure cell sample count is updated
If the number of samples to be collected is changed from one collection
to the next the data shape changes and the datawriter fails with an
'invalid frame' warning.

The new frame size is now checked when the number of samples is set and,
if different to the number of frames the file writer is expecting, a
'dummy scan' (where no file is written and no pressure jump is
triggered) is run to ensure the filewriter is expecting the correct
data.

Change-Id: Ide2bf6ded8ec7172674fe0fb781a215c6fe0c7c0