Use standard string formatting for scan complete message
When MessageFormat is used, scan numbers are formatted with commas as
thousands separators, eg `scan 123,456 completed successfully`. This is
not useful when searching for a scan number or when copying for use
elsewhere. This keeps the message the same but formats the number as it
used for filenames etc (eg, 123456).

Change-Id: I8793607fa894d57cd6b73643747a6f3c2ee5cde0