I14-1009 Add scan bean to DeviceController
The methods 'resume' and 'pause' in DeviceController attempt to set the
bean message with the watchdog message but the getBean method returns
null because it has not been set in the DeviceController. This change
fixes this and the scan will now display the reason why it was paused or
cannot resume, eg Beam has been lost.

Change-Id: I365983c473743476ebf9f897d449073b2da5ef2c