[DAQ-4615] remove fieldType field from MetadataStringItem record
This values is supposed to be set according to the type of the field,
i.e. was it an input or extra field. But in fact, this never actually
worked. The reason is that the value 'input' or 'extra' was put into a
NexusGroupData object, which was then converted back to a string
using ngdFieldType.toString(). Strangely, for both cases, this produces:
<dimensions><dimension>1</dimension></dimensions><type>NX_CHAR</type>.

Fixing this so that input names do come before extra fields causes the
tests to break. Instead, I decided to stick with the existing order,
and therefore all the code related to field type.

Change-Id: I3070da58dcf17d346f6456c266a88e51c9b5873e