B18-528 : Allow B18 'Sample parameter' GUI content to be customised
The order and grouping of the 'sample stage' motors in the B18 sample
parameters GUI can now be customised. This is controlled by a SampleMotorViewConfig
object configured in the client side spring.
Motors can be displayed together in logical groups (inside their own expandable
composite), and added to a specific column in the layout. The temperature
and sample wheel controls now appear in the same column, after the column(s)
of sample parameter motors.

Change-Id: I1d1761e836c80e2892adb66a7b6f71872d2b2941