DAQ-4692 Delete ScanPointGeneratorFactory init method
The method was only invoked from test code where it would not provide
any real benefit.

It was also observed to cause a deadlock in the tests on Jenkins.

At runtime in the application, it would conflict with GDA's other Jython
support if this was enabled for GDA which it isn't as noted in the
comment.

Change-Id: I6627003ef2f48c37f668923dee10ea64b67ef8b1