K11-1876 Ensure correct visit dir is read
AcquisitionFileBaseContext is backed by a simple map, which is populated
on first access and never changes. The map values are URLs which may
contain dirs which should change, such as visit.

In this commit, the values are changed to Supplier<URL> in order for the
URLs to be reevaluated whenever they are needed.

Change-Id: I762797ba2fdc703159f00ab754e58066e1d10dbe