I08-1 Add section to submit polarisation scans
This change allows to set a script to run before a scan. The script
request has a new boolean field so that the script that runs is
blocking, can be interrupted and throws exception. If an exception
occurs in the thread, a boolean flag terminated will be false, then the
scan that was going to be submitted after won't be submitted.

The polarisation values set in the ui will be placed in the Jython
namespace. Then, depending on the polarisation setting, the script will
caput different values to pvs and will also move the phase accordingly.

The change also includes a way to submit a scan with a predefined scan
request, instead of transforming the mapping bean in a scan request in
the ScanManagementController.

Change-Id: I3b2d56ba4e9c7b71c517d168d8ea2ad5a9ac05b1