DAQ-4731 Remove timeout in MJPEG reading
Set the url timeout to zero causing it to block indefinitely in the case
that there are no current frames, e.g. live stream stopped.

Previously this was handled by setting a fixed timeout and catching the
SocketTimeoutException within the loop. Changing this avoids using
exceptions for flow control.

In order to allow the read to be cancelled, it is wrapped in a
interruptableInputStream which will close the socket when the thread is
interrupted.

Change-Id: I143df3248602b6ef9afba82ae498831376565d8e