DAQ-4732 : Remove placeInJythonNamespace from SpectroscopyScanFactory
Placing the Jython object in the namespace automatically in calls to
'createScan' can lead to confusing behaviour if the 'scan name'
field doesn't match the name of the variable being assigned to in Jython.
e.g. the same object might referenced by two variables in the
Jython namespace; or subsequent calls to 'createScan' overwrite
previously created objects already in the namespace.

Change-Id: I62f2479d13e31c3e890c3c8ade600c395ba330e9