Replace plan setup wizard with simple dialog
The wizard was buggy, difficult to maintain, and difficult to use for
inexperienced users. The two tables showing segments and triggers were
visually counter-intuitive. These have now been replaced with a tree
viewer which more clearly shows the relationship between the components.
Creation/deletion of plan components are accessed via right-click menu,
which each option enabled/disabled depending on the current selection.
There is a single bit of context UI which is recreated/updated as the
components in the tree are selected.

ExperimentDriver was refactored slightly as part of this change:
 - no longer generic (only one model so far)
 - 'readouts' was confusing and had RMI serialisation issues when
   scannables where being accessed from the client. Now they are
   replaced by a list of DriverSignal records.

Note: the driver abstraction probably still needs heavy revision

The driver dialog can now be accessed through the plan setup dialog, and
the driver/profile selections of each are sensibly integrated.

Other changes:
 - PlanRequestParser now uses PlanCreator API
 - offsets in repeating triggers can now be defined in UI
 - improved the reset in PlanPreviewer (still needs improvements!)
 - added missing hashCode/equals methods to Trajectory and Scanpath-
    Document to fix selection issues in JFace viewers
 - fixed inconsistencies in subscriber logic which prevented reliably
   aborting running plans

Change-Id: I01a07b9d12fafa1d2e6cfa7a2ad85be49b47fc8b